JMS: cpptraj
The BOX_DIMENSIONS section of the topology file is not really used
anymore, even in explicit solvent simulations, since the box information
is typically stored in the INPCRD/RESTRT file. For this reason, I think
it's reasonable to support solvated/periodic prmtops without a %FLAG
BOX_DIMENSIONS. Thus, if that section is not present, just fill out that
information from the value of IFBOX (ifbox==0 --> no box, 1-->orthogonal,
and 2-->nonorthogonal/truncated octahedron).

Also add a 'chamber' boolean attribute to the AmberParm object so we
know when it's chamber or not. CHAMBER doesn't print BOX_DIMENSIONS
(which was the main motivation behind this change), so only print a
warning that they're missing if this is not a chamber prmtop.