JMS: cpptraj:
Bring CpptrajDevelopmentNotes.lyx into agreement with current code base
(it was quite out-of-date). Sections still need to be added about adding
Actions and Analyses (the two most common additions one would make, I'd
wager), but it should be accurate (hopefully; for the most part, anyway)
now.

Also fix a harmless typo and reduce an intra-sentence double-space to a
single space.