Fix a couple memory leaks in DCD file reading/writing that I introduced to make Dan look bad.
Allocate the boxtmp array on the stack rather than allocating in the
heap so it will automatically deallocate when it goes out of scope
(pointers allocated via new or malloc don't deallocate when their
reference goes out of scope, which I didn't realize).  This lead to a
leak of 6 doubles for every DCD frame that was read or written.

Since this is in bugfix.10 for AT 12, should this become another bugfix?