Add LIE (linear interaction energy) command to cpptraj. This command takes a "ligand" mask, a "surroundings" mask, and calculates the interactions between all atoms in the ligand mask with all atoms in the surroundings mask with a simple cutoff (tunable for both VDW and ELEC calcs). The minimum image convention is used, though a max of 1 image per particle is included.  LIE is intended to be done in explicit solvent with ligand-ligand and surroundings-surroundings interactions not included.
I also removed the requirement that the bonds and bondsh arrays be populated,
since you can have a system of monoatomic ions that have no bonds. (I used this
as a simple test of LIE). Still emit a warning, but allow Amber topologies with
no bonds.

Thanks to Dan for helping me find and fix a segfault.