JMS - Cpptraj: Add a volmap command that uses the same approach as the VMD VolMap tool to calculate particle densities (i.e., it treats each particle as a normalized Gaussian of width <radius> and distributes particle density on a the grid for each particle).
Added a couple functions in the Grid class to accommodate the operations I
needed to perform.

  - Fixed the 'help' in "grid", where the command name was replaced by a tab
  - Fixed an error message in Grid where the string-replacements were swapped

Added test files for the new action. The density maps it prints out superimpose
almost perfectly on the VMD-generated maps if the radii are scaled to match what
VMD uses. I extract the radii from the VDW A- and B-coefficients, using 0.5*Rmin
for the diagonal terms.