Adjust the external NetCDF option to use dynamic linking (unless -static is explicitly requested). As a result, I had to remove the numerous checks that $(NETCDFLIB[F]) existed, since $(NETCDFLIB[F]) always evaluated to "-L/path/to/netcdf/lib" instead of "/path/to/netcdf/lib/libnetcdf.a".
I did this ultimately to provide a pathway to supporting files created with
NetCDF-4 because it is important.  Not because Amber uses these features but
because other packages that write NetCDF files that Amber would want to use
write files that cannot currently be understood. (e.g., OpenMM/mdtraj, any
Python script that uses the Python NetCDF bindings to create a trajectory, etc.)

Building the bundled NetCDF with NetCDF4 support is a bad idea due to the large
number of dependencies it introduces (HDF5, etc.). Furthermore, static linking
with libnetcdff.a is effectively broken -- it actually seems to depend on linked
libraries itself. This seems wrong. The only way I could get any of my package
manager NetCDF installations to work with Amber was to link dynamically.

As a result, the "--with-netcdf" flag triggers dynamic linking (the distro
NetCDF is still linked statically when possible to avoid the types of errors we
saw in the past). The existence checks that were performed in all of the
Makefiles were redundant -- the libraries and headers are checked inside
configure when test programs were compiled. Since it was too much work to get
them to work with external NetCDFs, I just did away with them.