Attempt to fix pytraj on OS X.
Building directly into $(LIBDIR) can be considered a temporary hack until we get
the install_name_tool working correctly. However, this has no effect on the
standalone configure script, and is useful for AmberTools since pytraj is
currently the only user of the cpptraj API.  This is needed to work toward
getting pytraj to work on Macs.