Try to fix permissions of pytraj building with this docker image inside Jenkins.
The problem is that Jenkins doesn't run as root, it runs as the UID of
the account that happens to be running the job. This is required for a
few reasons, but it also means that we need to make sure that *everyone*
has permission to write to the directories that conda needs to write to.
Otherwise all of the packages have to be added here (meaning that the
requirements will leak from pytraj into this file).