Remove the Intel compiler build (#895)
I don't want to do this, but Intel makes it harder and harder each year to renew the license (which has currently expired).  I'll leave the block in there in case Intel makes it easier or possible to do this in the future, but let's not block pytraj testing because my Intel compiler license has expired and I can't bother spending the hour(s) I need to try and fix it.