Speed up six-dimensional entropy calculation in GIST
In GIST, the six-dimensional entropy calculation requires the
translational and orientational distance to the nearest neighbor (NN).
However, if the translational distance is larger than the 6D distance to
the current NN, the orientational distance can be skipped safely.  This
is a significant speedup (about 2x in my test), since the orientational
distance contains an arccos.