Implement exact volume equations for GIST nearest-neighbor entropy.
The nearest-neighbor density is approximated by finding the distance to
the nearest neighbor of each point, computing a sphere volume from this
distance, and comparing this to the volume that is expected from a
homogeneous density. In GIST, the sphere volume in rotational space and
the combined translation+rotation space was approximated by a Taylor
expansion. This patch implements the exact volume in rotational space
analytically, and uses a tabulated correction factor for distances up to
10 Angstrom to correct the 6-D volume.

The result is a better convergence of the bulk entropy, at the cost of
about 3.4% calculation time (Action Post) in my test (a pure water box
containg 3333 water molecules, using 10000 frames). The bulk entropy
(should be zero) is reduced from 2.3e-4 to 5e-5 kcal/mol/A^3 for dTSsix
and from 7.7e-4 to 2.3e-4 for dTSorient. The individual voxel values are
almost perfectly correlated to the old results (R^2=0.99996).