Implement generalized nearest neighbor (NN) search.
In the old algorithm, neighbors were searched in the respective voxel
and in all of its direct neighbors. The new algorithm lets the user
choose how many layers of neighboring voxels should be search. The
default is 1, which corresponds to the old behavior.
There is a tiny difference in how distances of 0 are treated. In the old
code, they were omitted. In the new code, there is a constant
(GIST_TINY) which is used if the resulting distance is 0. This leads to
numerically different values in about 0.01% of the voxels.