In GIST, omit entropy for waters if no nearest neighbor is found.
Changes:
* after the nearest neighbor (NN) search, introduces a check whether the
  distance is smaller than sqrt(GIST_HUGE) (i.e., a NN has been found). If
  not, this distance is omitted from the entropy calculation.
* fix a bug where some non-border voxels were omitted instead of voxels
  on the border. (GIST voxels on the border of the grid are supposed to
  obtain 0 entropy)
* don't look for nearest neighbors if we are on the border.

Explanation:
In GIST, the entropy is calculated per water molecule based on the
NN distance, and binned to a grid. With the recent change of the entropy
calculation, I removed some checks that:
* omit the entropy of waters if no NN distance is found. Instead, a high
  value (sqrt(GIST_HUGE)) is assigned to the distance.
* omit the NN if the distance to it is numerically close to zero.
  Instead, a small value (GIST_TINY) is then assigned to the distance.
The second first behavior is reasonable (omitting the NN means
that an arbitrary higher distance is chosen, and GIST_TINY is probably
smaller than that arbitrary distance, i.e., closer to the real one.
However, the first one is not so reasonable, because sqrt(GIST_HUGE) is
probably far off from the real NN distance. Therefore, I re-introduce
the old behavior in this case.

This is very relevant for the results of the cpptraj tests, since they
use very few frames, leading to many cases of high NN distances.