In GIST, track molecular densities and set rigidatomindices
rigidAtomIndices_ specifies which atoms of the solvent define a rigid
subunit. This is needed for the entropy calc, and also (if "nocom" is
specified) to define the molecule's position. Each number is an offset
from the molecule's start index (i.e., 0 means the first atom of the
moleucle). The first number is the central atom, and the second and
third number are other atoms that are bound to it. It can be set in 2
ways:
* The user specifies rigidatoms. In that case, the atom names are
  matched to set the indices (offset from molecule start)
* If the user doesn't specify, the solvent atom with the most bonds is
  searched. If it has less than 2 bonds to heavy atoms, hydrogens are
  also included. The central atom is the one with the most bonds. The
  two others are chosen from the atoms that are bonded to it.

Molecular densities are tracked in separate datasets. They are defined
relative to the solvent bulk density, and are computed by binning the
solvent centers of each specified solvent species (either the COM or the
first of "rigidAtomIndices_"). The user has to choose which molecules
are tracked by specifying "solventmols", followed by a comma-separated
list of molecule names. No selection masks are allowed, firstly because
this would mess with the comma-separated list, and also because we
anyway assume that we select whole molecules. For each specified
molecule, a dataset is generated and written out as dx file, e.g.,
gist-g-mol-WAT.dx