Fix compilation on ARM based systems (#1093)
* Handle case where 0 is passed to FloatWidth()

* Change FloatWidth() function to MinNumFracChars(). FloatWidth() was a
misleading name; the original intent was to ensure when writing out
coordinates with a certain step size that the precision was large enough
to accomodate the step (e.g. a step of .001 would need at least a
precision of 3). Also rework the function so it is more accurate for
values that are not exactly powers of 10.

* 6.28.0. Minor revision bump for fixes for ARM; may also slightly affect
  (fix) the output precision of certain X step values in output ASCII
data formats in some cases.